structuring viewmodel classes Separated functions for fragments into their respective viewmodels
