+++ app/src/main/java/com/judahben149/noteify/viewmodel/DeletedNoteViewModel.kt
@@ -6,78 +6,50 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.viewModelScope
 import com.judahben149.noteify.data.NoteDatabase
 import com.judahben149.noteify.model.Note
-import com.judahben149.noteify.model.PrivateNote
 import com.judahben149.noteify.repository.NoteRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 class DeletedNoteViewModel(application: Application) : AndroidViewModel(application) {
 
-        val readAllNotes: LiveData<List<Note>>
-        val readAllFavoriteNote: LiveData<List<Note>>
-        val readAllDeletedNotes: LiveData<List<Note>>
-        val readAllPrivateNotes: LiveData<List<PrivateNote>>
-        private val repository: NoteRepository
-
-//    val isNoteFavorite: Boolean = false
-
-        init {
-            val noteDao = NoteDatabase.getDatabase(application).noteDao()
-            repository = NoteRepository(noteDao)
-            readAllNotes = repository.readAllNotes
-            readAllFavoriteNote = repository.readAllFavoriteNotes
-            readAllDeletedNotes = repository.readAllDeletedNotes
-            readAllPrivateNotes = repository.readAllPrivateNotes
-        }
-
-        fun addNote(note: Note) {
-            viewModelScope.launch(Dispatchers.IO) {
-                repository.addNote(note)
-            }
-        }
-
-        fun updateNote(note: Note) {
-            viewModelScope.launch(Dispatchers.IO) {
-                repository.updateNote(note)
-            }
-        }
-
-        fun deleteNote(note: Note) {
-            viewModelScope.launch(Dispatchers.IO) {
-                repository.deleteNote(note)
-            }
-        }
-
-        fun deleteAllNotes() {
-            viewModelScope.launch(Dispatchers.IO) {
-                repository.deleteAllNotes()
-            }
-        }
+    val readAllDeletedNotes: LiveData<List<Note>>
+    private val repository: NoteRepository
 
 
-        fun searchDatabase(searchQuery: String): LiveData<List<Note>> {
-            return  repository.searchDatabase(searchQuery)
-        }
-
-
-
-        //methods for private notes
-        fun addPrivateNote(note: PrivateNote) {
-            viewModelScope.launch(Dispatchers.IO) {
-                repository.addPrivateNote(note)
-            }
-        }
-
-        fun deletePrivateNote(note: PrivateNote) {
-            viewModelScope.launch(Dispatchers.IO) {
-                repository.deletePrivateNote(note)
-            }
-        }
-
-        fun deleteAllPrivateNotes() {
-            viewModelScope.launch(Dispatchers.IO) {
-                repository.deleteAllPrivateNotes()
-            }
-        }
-
+    init {
+        val noteDao = NoteDatabase.getDatabase(application).noteDao()
+        repository = NoteRepository(noteDao)
+        readAllDeletedNotes = repository.readAllDeletedNotes
     }
+
+    //methods for deleted notes
+    fun updateNote(note: Note) {
+        viewModelScope.launch(Dispatchers.IO) {
+            repository.updateNote(note)
+        }
+    }
+
+    fun deleteNote(note: Note) {
+        viewModelScope.launch(Dispatchers.IO) {
+            repository.deleteNote(note)
+        }
+    }
+
+    fun deleteAllDeletedNotes() {
+        viewModelScope.launch(Dispatchers.IO) {
+            repository.deleteAllDeletedNotes()
+        }
+    }
+
+//    fun deleteAllNotes() {
+//        viewModelScope.launch(Dispatchers.IO) {
+//            repository.deleteAllNotes()
+//        }
+//    }
+
+    fun restoreNotesFromTrash() {
+        viewModelScope.launch(Dispatchers.IO) {
+            repository.restoreNotesFromTrash()
+        }
+    }
+}
